Move property_behavior tests into "Prototypes" dir.
The "lazy" and "delayed" test cases reflect things we could conceivably bring into the standard library at some point, and the separate files make it more clear to experimenters how the partially-implemented bits of the feature manage to work.
